[IB]fixed the limit of starter memory size (2GB).
git-svn-id: svn://acecls.cc.kyushu-u.ac.jp/ACE/devel@285 d6456ff6-7713-4bd9-81c0-de8efccfd6a1

+++ src/bl/ib/acpbl_ib.c
@@ -4076,7 +4076,7 @@ int iacp_init(void){
     int flags; /* flag of modify queue pair */
     
     /* allocate the starter memory adn register memory */
-    int syssize;
+    size_t syssize;
     /* ajusting variable for alignment 8 bytes */
     int alm8_add; 
     /* enable address resuse, as soon as possible */
